Update Trino to 405 and add support for query pass-through
Updated Trino version to 405.

Added support for query pass-through from Trino to the connected DB2 instance. See https://github.com/trinodb/trino/pull/12325 for more details about the implementation of the ``query`` Polymorphic Table Function for JDBC-based connectors in Trino.

The functionality is useful when dealing with very large tables on DB2 as Trino does not need to pull entire tables across. It also allows the use of DB2-native functions that are not supported in Trino.